haha, ok - so this is a bit funny
It looks like - the changes that I have made are not impacting the tests
at all. Like - the most recent changes with the lists (which is - fair).

However - by adding an additional 'parsing' argument - the regexes
started breaking.

This means to me - that I probably need to replace the "dot" char into
a proper - char for lists.